Rewrite so EventListener invokes global function
The way the code was written prior to this didn't allow for the user to change the drawing
color without everything being erased.

With this rewrite the EventListener calls a function that is defined outside of the EventListener.
Meaning its functionality can be updated without having to redefine every instance of the EventListener.